Fix SSR crash in ShareButtons and split DeckPageClient into UI/logic
ShareButtons read window.location.origin unconditionally during render
whenever isLocal was true, crashing SSR (500) on any /deck/local/... page
such as the one right after creating a new deck. Guarded with typeof
window check.

Extracted DeckPageClient's Firebase/localStorage/deck-editing logic into
useDeckPageState so the component only assembles JSX. Consolidated the
tab-switching + per-tab coloring + card list, which used to be split
between DeckPageClient and CardList, into a single CardListPanel.
CardList itself now delegates filtering/sorting to useCardListFilters
and its filter UI to CardListFilterBar.

Also removed a few unused imports/props flagged by lint (HeroSection's
unused isMounted prop, unused type imports in framer-motion.d.ts).
