+++ src/app/page.tsx
@@ -28,7 +28,7 @@ export default function Home() {
   return (
     <main className={css({ minH: 'screen', w: 'full', position: 'relative', overflowX: 'hidden' })}>
       {/* ヒーローセクション */}
-      <HeroSection cardsYPosition={cardsYPosition} isMounted={isMounted} />
+      <HeroSection cardsYPosition={cardsYPosition} />
 
       {/* 解説セクション */}
       {isMounted && (
